requirements.txt
openai>=1.0.0
rich>=13.0.0
ddgs>=1.0.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
requests>=2.31.0
prompt_toolkit>=3.0.0
